+++ src/cmd/config.rs
@@ -1,7 +1,7 @@
 use crate::config::{default_config_path, models::ConfigOpts};
 use crate::inventory::default_inventory_path;
-use crate::io;
-use crate::io::scanner::{prompt, prompt_ron, prompt_t, prompt_yes_or_no};
+use crate::util;
+use crate::util::scanner::{prompt, prompt_ron, prompt_t, prompt_yes_or_no};
 use anyhow::Result;
 use clap::{Args, Subcommand};
 use std::path::PathBuf;
@@ -137,7 +137,7 @@ async fn build() -> Result<()> {
             })
             .unwrap_or(default_path)
     };
-    io::fs::remove_interactive(&path).await?;
+    util::fs::remove_interactive(&path).await?;
     builder.save(path).await?;
 
     Ok(())
