+++ src/config/mod.rs
@@ -1,6 +1,6 @@
-//! CDDNS config.
+//! cddns configuration.
 //!
-//! CDDNS takes the typical layered configuration approach. There are 3 layers.
+//! cddns takes the typical layered configuration approach. There are 3 layers.
 //! The config file is the base, which is then superseded by environment
 //! variables, which are finally superseded by CLI arguments and options.
