+++ src/util/encoding.rs
